fix: move locales files to src
