+++ package.json
@@ -144,6 +144,7 @@
     "vite": "6.0.1",
     "vite-plugin-checker": "0.8.0",
     "vite-plugin-istanbul": "6.0.2",
+    "vite-plugin-static-copy": "2.2.0",
     "vitest": "2.1.6"
   },
   "volta": {
