+++ src/locales/ar/player.json
