+++ src/locales/de/enums.json
