+++ src/locales/de/landing.json
