+++ src/locales/de/player.json
