+++ src/locales/en/player.json
