+++ src/locales/es/player.json
