+++ src/locales/fr/account.json
