+++ src/locales/fr/enums.json
