+++ src/locales/fr/landing.json
