+++ src/locales/fr/player.json
