+++ src/locales/it/player.json
