+++ src/modules/auth/components/signIn/EmailInput.tsx
@@ -1,7 +1,4 @@
 import { UseFormRegisterReturn } from 'react-hook-form';
-import { useTranslation } from 'react-i18next';
-
-import { NS } from '@/config/constants';
 
 import { StyledTextField } from '../common/StyledTextField';
 import { EmailAdornment } from '../common/adornments';
@@ -23,8 +20,6 @@ export function EmailInput({
   error,
   placeholder,
 }: Readonly<Props>): JSX.Element {
-  const { t } = useTranslation(NS.Auth);
-
   return (
     <StyledTextField
       id={id}
@@ -36,12 +31,7 @@ export function EmailInput({
       variant="outlined"
       error={Boolean(error)}
       placeholder={placeholder}
-      helperText={
-        error &&
-        // eslint-disable-next-line @typescript-eslint/ban-ts-comment
-        // @ts-expect-error
-        t(error)
-      }
+      helperText={error}
       // eslint-disable-next-line jsx-a11y/no-autofocus
       autoFocus={autoFocus}
       disabled={disabled}
